requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
factory_boy==3.3.3
Faker==37.0.2
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.3
tzdata==2025.1
